fix: remove tauri::ipc::Channel type from generated type set
After collecting types from all the commands, make sure to remove the
type for `tauri::ipc::Channel` as it may be present if any command took
the channel as argument.

Until this addition the user would generate the TAURI_CHANNEL type in
their bindings if they called `.events()` and then `.commands()` in this
order.

The call to `.events()` tried to remove this type, but it got added
later in the call to `.commands()`.